Fix synchronization issues for the players list.
There we still issues and race conditions, most notably seen when
the server restarted.  The /rooms/members/1 JSON file is no longer
used in tandem with the websocket.  It is used once at page load
just to ensure browsers without websocket support can still view
the site.

For the rest, there now is a room_list command that shows the users
in a room when requested.  This is the first command sent on a socket
when the room has been joined.

Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>